.docker/python_middleware/requirements.txt
fastapi
uvicorn
python-dotenv
requests
python-jose[cryptography]
cachetools
pydantic-settings